fix some warning from mypy

+++ tartare/core/models.py
@@ -163,7 +163,7 @@ class PreProcessContainer(metaclass=ABCMeta):
                     self.data_sources.append(data_source_computed)
 
     @classmethod
-    def get(cls, object_id: str) -> Union['Contributor', 'Coverage']:
+    def get(cls, object_id: str) -> Union['Contributor', Optional['Coverage']]:
         pass
 
 
@@ -187,7 +187,7 @@ class Platform(object):
 
 
 class PublicationPlatform(SequenceContainer, Platform):
-    def __init__(self, protocol: str, type: str, url: str, options: PlatformOptions = None, sequence: Optional[int] = 0,
+    def __init__(self, protocol: str, type: str, url: str, options: PlatformOptions = None, sequence: int = 0,
                  input_data_source_ids: List[str] = None) -> None:
         SequenceContainer.__init__(self, sequence)
         Platform.__init__(self, protocol, url, options)
@@ -198,7 +198,7 @@ class PublicationPlatform(SequenceContainer, Platform):
 class Environment(SequenceContainer):
     def __init__(self, name: str = None, current_ntfs_id: str = None,
                  publication_platforms: List[PublicationPlatform] = None,
-                 sequence: Optional[int] = 0) -> None:
+                 sequence: int = 0) -> None:
         super().__init__(sequence)
         self.name = name
         self.current_ntfs_id = current_ntfs_id
@@ -294,7 +294,7 @@ class DataSet(object):
         self.validity_period = validity_period
         self.status_history = status_history if status_history else []
 
-    def get_md5(self) -> str:
+    def get_md5(self) -> Optional[str]:
         if not self.gridfs_id:
             return None
         file = GridFsHandler().get_file_from_gridfs(self.gridfs_id)
@@ -317,8 +317,8 @@ class DataSet(object):
 class DataSource(object):
     def __init__(self, id: Optional[str] = None,
                  name: Optional[str] = None,
-                 data_format: Optional[str] = DATA_FORMAT_DEFAULT,
-                 input: Optional[Input] = Input(INPUT_TYPE_DEFAULT),
+                 data_format: str = DATA_FORMAT_DEFAULT,
+                 input: Input = Input(INPUT_TYPE_DEFAULT),
                  license: Optional[License] = None, export_data_source_id: str = None,
                  service_id: str = None, data_sets: List[DataSet] = None) -> None:
         self.id = id if id else str(uuid.uuid4())
@@ -370,7 +370,7 @@ class DataSource(object):
         return data_sources[0]
 
     @classmethod
-    def get_contributor_of_data_source(cls, data_source_id: str) -> 'Contributor':
+    def get_contributor_of_data_source(cls, data_source_id: str) -> Optional['Contributor']:
         raw = mongo.db[Contributor.mongo_collection].find_one({'data_sources': {'$elemMatch': {'id': data_source_id}}})
 
         if not raw:
@@ -390,7 +390,7 @@ class DataSource(object):
         return nb_delete
 
     @classmethod
-    def update(cls, contributor_id: str, data_source_id: str = None, dataset: dict = None) -> 'DataSource':
+    def update(cls, contributor_id: str, data_source_id: str = None, dataset: dict = None) -> Optional['DataSource']:
         tmp_dataset = dataset if dataset else {}
         if data_source_id is None:
             raise ValueError('a data_source id is required')
@@ -403,12 +403,11 @@ class DataSource(object):
 
         # `$` acts as a placeholder of the first match in the list
         contrib_dataset = {'data_sources': {'$': tmp_dataset}}
-        raw = mongo.db[Contributor.mongo_collection].update_one({'data_sources.id': data_source_id},
-                                                                {'$set': to_doted_notation(contrib_dataset)})
-        if raw.matched_count == 0:
-            return None
+        mongo.db[Contributor.mongo_collection].update_one({'data_sources.id': data_source_id},
+                                                          {'$set': to_doted_notation(contrib_dataset)})
 
-        return cls.get(contributor_id, data_source_id)[0]
+        data_sources = cls.get(contributor_id, data_source_id)
+        return data_sources[0] if data_sources else None
 
     def add_data_set_and_update_model(self, data_set: DataSet, model: Union['Contributor', 'Coverage']) -> None:
         self.data_sets.append(data_set)
@@ -416,7 +415,8 @@ class DataSource(object):
         if len(self.data_sets) > data_sets_number:
             sorted_data_set = sorted(self.data_sets, key=lambda ds: ds.created_at, reverse=True)
             for data_set_to_be_removed in sorted_data_set[data_sets_number:]:
-                GridFsHandler().delete_file_from_gridfs(data_set_to_be_removed.gridfs_id)
+                if data_set_to_be_removed.gridfs_id:
+                    GridFsHandler().delete_file_from_gridfs(data_set_to_be_removed.gridfs_id)
             self.data_sets = sorted_data_set[0:data_sets_number]
         model.update()
 
@@ -430,12 +430,12 @@ class DataSource(object):
         return self.input.type == type
 
     def get_last_data_set(self) -> Optional[DataSet]:
-        return max(self.data_sets, key=lambda ds: ds.created_at, default=None)
+        return max(self.data_sets, key=lambda ds: ds.created_at, default=None)  # type: ignore
 
 
 class GenericPreProcess(SequenceContainer):
     def __init__(self, id: Optional[str] = None, type: Optional[str] = None, params: Optional[dict] = None,
-                 sequence: Optional[int] = 0, data_source_ids: Optional[List[str]] = None,
+                 sequence: int = 0, data_source_ids: Optional[List[str]] = None,
                  enabled: bool = True) -> None:
         super().__init__(sequence)
         self.id = str(uuid.uuid4()) if not id else id
@@ -498,7 +498,7 @@ class GenericPreProcess(SequenceContainer):
     @classmethod
     def update_data(cls, class_name: Type[PreProcessContainer],
                     mongo_schema: Type['MongoPreProcessContainerSchema'], object_id: str,
-                    preprocess_id: str, preprocess: Optional[dict] = None) -> Optional[List['PreProcess']]:
+                    preprocess_id: str, preprocess: Optional[Dict[str, Any]] = None) -> Optional[List['PreProcess']]:
         data = class_name.get(object_id)
         if not data:
             raise ValueError('bad {} {}'.format(class_name.label, object_id))
@@ -506,7 +506,7 @@ class GenericPreProcess(SequenceContainer):
         if not [ps for ps in data.preprocesses if ps.id == preprocess_id]:
             raise ValueError("no preprocesses id {} exists in {} with id {}"
                              .format(object_id, class_name.label, preprocess_id))
-        if 'id' in preprocess and preprocess['id'] != preprocess_id:
+        if preprocess and 'id' in preprocess and preprocess['id'] != preprocess_id:
             raise ValueError("id from request {} doesn't match id from url {}"
                              .format(preprocess['id'], preprocess_id))
 
@@ -531,7 +531,7 @@ class PreProcess(GenericPreProcess):
             self.save_data(Coverage, MongoCoverageSchema, coverage_id, self)
 
     @classmethod
-    def get(cls, preprocess_id: Optional[str] = None, contributor_id: Optional[str] = None,
+    def get(cls, preprocess_id: str, contributor_id: Optional[str] = None,
             coverage_id: Optional[str] = None) -> Optional[List['PreProcess']]:
         if not any([coverage_id, contributor_id]):
             raise ValueError('bad arguments')
@@ -611,15 +611,6 @@ class Contributor(PreProcessContainer):
     def all(cls) -> List['Contributor']:
         return cls.find(filter={})
 
-    @classmethod
-    def update_with_dict(cls, contributor_id: str = None, contributor_dict: dict = None) -> Optional['Contributor']:
-        contributor_dict = contributor_dict if contributor_dict else {}
-        raw = mongo.db[cls.mongo_collection].update_one({'_id': contributor_id}, {'$set': contributor_dict})
-        if raw.matched_count == 0:
-            return None
-
-        return cls.get(contributor_id)
-
     def update(self) -> None:
         self.update_with_object(self)
 
@@ -663,13 +654,11 @@ class Coverage(PreProcessContainer):
         Coverage.update_with_dict(self.id, {'grid_calendars_id': id})
         # when we delete the file all process reading it will get invalid data
         # TODO: We will need to implement a better solution
-        gridfs_handler.delete_file_from_gridfs(self.grid_calendars_id)
+        if self.grid_calendars_id:
+            gridfs_handler.delete_file_from_gridfs(self.grid_calendars_id)
         self.grid_calendars_id = id
 
-    def get_environment(self, environment_id: str) -> Environment:
-        return self.environments.get(environment_id)
-
-    def get_grid_calendars(self) -> str:
+    def get_grid_calendars(self) -> Optional[GridOut]:
         if not self.grid_calendars_id:
             return None
 
@@ -681,7 +670,7 @@ class Coverage(PreProcessContainer):
         mongo.db[self.mongo_collection].insert_one(raw)
 
     @classmethod
-    def get(cls, coverage_id: str = None) -> 'Coverage':
+    def get(cls, coverage_id: str = None) -> Optional['Coverage']:
         raw = mongo.db[cls.mongo_collection].find_one({'_id': coverage_id})
         if raw is None:
             return None
@@ -703,19 +692,18 @@ class Coverage(PreProcessContainer):
         return cls.find(filter={})
 
     @classmethod
-    def update_with_dict(cls, coverage_id: str = None, dataset: dict = None) -> 'Coverage':
+    def update_with_dict(cls, coverage_id: str = None, dataset: dict = None) -> None:
         # we have to use "doted notation' to only update some fields of a nested object
         tmp_dataset = dataset if dataset else {}
-        raw = mongo.db[cls.mongo_collection].update_one({'_id': coverage_id}, {'$set': to_doted_notation(tmp_dataset)})
-        if raw.matched_count == 0:
-            return None
+        mongo.db[cls.mongo_collection].update_one({'_id': coverage_id}, {'$set': to_doted_notation(tmp_dataset)})
 
-        return cls.get(coverage_id)
+        cls.get(coverage_id)
 
     def __fill_password_from_existing_coverage(self, existing_coverage: 'Coverage') -> None:
         for env_key, environment in self.environments.items():
             for platform in environment.publication_platforms:
-                if platform.options and platform.options.authent and not platform.options.authent.password:
+                if platform.options and platform.options.authent and platform.options.authent.username and \
+                        not platform.options.authent.password:
                     if env_key in existing_coverage.environments:
                         existing_platform = existing_coverage.environments[env_key] \
                             .get_publication_platform_for_type_with_user(
@@ -764,8 +752,8 @@ class Coverage(PreProcessContainer):
             self.contributors_ids.remove(contributor_id)
             self.update_with_dict(self.id, {"contributors_ids": self.contributors_ids})
 
-    def get_data_source(self, data_source_id: str) -> Optional['DataSource']:
-        return next((data_source for data_source in self.data_sources if data_source.id == data_source_id), None)
+    def get_data_source(self, data_source_id: str) -> 'DataSource':
+        return next(data_source for data_source in self.data_sources if data_source.id == data_source_id)
 
     def __repr__(self) -> str:
         return str(vars(self))
@@ -906,8 +894,8 @@ class Historisable(object):
             # Delete old data_sources_fetched
             num_deleted = self.delete_many([row.get('_id') for row in old_rows])
             # Delete all associated gridFS
+            gridfs_ids = []  # type: List[str]
             if num_deleted:
-                gridfs_ids = []  # type: List[str]
                 get_values_by_key(old_rows, gridfs_ids)
             for gridf_id in gridfs_ids:
                 GridFsHandler().delete_file_from_gridfs(gridf_id)
@@ -1205,8 +1193,6 @@ class CoverageExport(Historisable):
 
     @classmethod
     def get_last(cls, coverage_id: str) -> Optional['CoverageExport']:
-        if not coverage_id:
-            return None
         raw = mongo.db[cls.mongo_collection].find({'coverage_id': coverage_id}).sort("created_at", -1).limit(1)
         lasts = MongoCoverageExportSchema(many=True).load(raw).data
         return lasts[0] if lasts else None
